+++ core/Hierarchy/EntityProfiles/SensorProfiles/eoSensorLib/eovision.h
@@ -5,11 +5,11 @@ class EOVision
 {
 public:
     EOVision();
-    Vec3 geoToECEF(double lat, double lon, double alt);
-    Vec3 ecefToENU(Vec3 target, Vec3 sensor, double lat_s, double lon_s);
-    Vec3 rotateToSensorFrame(Vec3 v, double headingDeg, double pitchDeg);
-    bool isInsideFOV(Vec3 v, double hfovDeg, double vfovDeg);
-    Vec2 project(Vec3 v, double focalLength);
+    Vec3 geoToECEF(float lat, float lon, float alt);
+    Vec3 ecefToENU(Vec3 target, Vec3 sensor, float lat_s, float lon_s);
+    Vec3 rotateToSensorFrame(Vec3 v, float headingDeg, float pitchDeg);
+    bool isInsideFOV(Vec3 v, float hfovDeg, float vfovDeg);
+    Vec2 project(Vec3 v, float focalLength);
 };
 
 #endif // EOVISION_H
